[REFACTOR] Rename mapper and restructure of internal classes.
