+++ smartadapter/src/main/java/smartadapter/SmartAdapterBuilder.java
@@ -17,7 +17,7 @@ import smartadapter.viewholder.SmartAutoEventViewHolder;
 import smartadapter.viewholder.SmartViewHolder;
 import smartadapter.widget.ViewTypeResolver;
 
-import static smartadapter.internal.ViewEventValidator.isViewEventIdValid;
+import static smartadapter.internal.utils.ViewEventValidator.isViewEventIdValid;
 
 /**
  * Builder for SmartRecyclerAdapter.
